Add files via upload
Wine quality prediction using machine learning (ML) by linear regression is a process of using ML algorithms to learn the relationship between a set of features, such as acidity, alcohol content, and pH, and the quality of wine. Once the algorithm has learned the relationship, it can be used to predict the quality of new wine samples.

Linear regression is a supervised learning algorithm, which means that it learns from a dataset of labeled examples, where the input features are known and the output target variable is the quality of the wine. The algorithm works by finding a linear function that best fits the data. This function can then be used to predict the quality of new wine samples by plugging in the values of the features.